Rename relevant files to use kebab-case
