+++ src/setupTests.ts
@@ -1,3 +1,7 @@
+// This file is not renamed to `setup-tests.ts` as this is only a default from Create React
+// App to use Jest as its tests. Rather than changing the defaults, I'd rather keep them
+// the same to prevent any unwanted side-effects.
+
 // jest-dom adds custom jest matchers for asserting on DOM nodes.
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
